requirement2.txt
Flask==2.3.3
Flask-Smorest==0.43.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
bcrypt==4.0.1
passlib==1.7.4
cryptography==42.0.5